pyproject.toml
[project]
name = "lisa"
dynamic = ["version"]
description = "Looping Implementor Saving Assumptions - Autonomous Linear ticket implementation with Claude"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Evenly", email = "dev@evenly.no" }
]
keywords = ["linear", "claude", "automation", "ai", "code-generation", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.urls]
Repository = "https://github.com/evenly-energy/lisa"
Issues = "https://github.com/evenly-energy/lisa/issues"
Changelog = "https://github.com/evenly-energy/lisa/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
    "pre-commit>=3.5",
    "time-machine>=2.13",
]

[project.scripts]
lisa = "lisa.cli:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/lisa"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]  # Line length handled by formatter

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
show_error_codes = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.coverage.run]
source = ["lisa"]
branch = true
omit = ["src/lisa/ui/assumptions.py"]

[tool.coverage.report]
fail_under = 70
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
